+++ src/classes/babies/SnailBaby.ts
@@ -1,12 +1,31 @@
-import { CacheFlag, ModCallback } from "isaac-typescript-definitions";
-import { Callback } from "isaacscript-common";
+import { CollectibleType } from "isaac-typescript-definitions";
+import {
+  CallbackCustom,
+  ModCallbackCustom,
+  useActiveItemTemp,
+} from "isaacscript-common";
 import { Baby } from "../Baby";
 
-/** 0.5x speed. */
+const v = {
+  room: {
+    numHits: 0,
+  },
+};
+
+/** Glowing Hourglass effect on Nth hit (per room). */
 export class SnailBaby extends Baby {
-  // 8
-  @Callback(ModCallback.EVALUATE_CACHE, CacheFlag.SPEED)
-  evaluateCacheSpeed(player: EntityPlayer): void {
-    player.MoveSpeed *= 0.5;
+  v = v;
+
+  @CallbackCustom(ModCallbackCustom.ENTITY_TAKE_DMG_PLAYER)
+  entityTakeDmgPlayer(player: EntityPlayer): boolean | undefined {
+    const num = this.getAttribute("num");
+
+    v.room.numHits++;
+    if (v.room.numHits >= num) {
+      v.room.numHits = 0;
+      useActiveItemTemp(player, CollectibleType.GLOWING_HOUR_GLASS);
+    }
+
+    return undefined;
   }
 }